last factories

+++ src/Item.java
@@ -1,7 +1,7 @@
 public abstract class Item {
     String name;
-    int cost;
-    int reqLevel;
+    double cost;
+    double reqLevel;
 
     public Item() {
         this.name = null;
@@ -9,7 +9,7 @@ public abstract class Item {
         this.reqLevel = 0;
     }
 
-    public Item(String name, int cost, int reqLevel) {
+    public Item(String name, double cost, double reqLevel) {
         this.name = name;
         this.cost = cost;
         this.reqLevel = reqLevel;
@@ -23,19 +23,19 @@ public abstract class Item {
         this.name = name;
     }
 
-    public int getCost() {
+    public double getCost() {
         return cost;
     }
 
-    public void setCost(int cost) {
+    public void setCost(double cost) {
         this.cost = cost;
     }
 
-    public int getReqLevel() {
+    public double getReqLevel() {
         return reqLevel;
     }
 
-    public void setReqLevel(int reqLevel) {
+    public void setReqLevel(double reqLevel) {
         this.reqLevel = reqLevel;
     }
